+++ CHANGELOG.md
@@ -1,8 +1,12 @@
 # 0.1.1
+- WIP (ainda não disponível para uso em produção)
 - Implementada consulta de NFEs(V1) por id do gateway
 - Implementada consulta de NFEs(V1) por id externo
 - Implementada pesquisa de NFEs(v1)
+- Implementado Incluir/Alterar Empresa
+- Implementado Habilitar Empresa
+- Implementado Desabilitar Empresa
 
 # 0.1.0
-- Versão em desenvolvimento, ainda não disponível para utilizar em produção
+- WIP (ainda não disponível para uso em produção)